==> 02-Data-sources.Rmd <==
# Data sources 

Our data sources are from first hand public data source platform. The Covid data is from the NYC Health github website; Crime data is from the NYC Open Data, the Police Department (NYPD). 


- [Link to Covid Vaccination Data](https://github.com/nychealth/coronavirus-data/tree/master/latest)

    - now-weekly-breakthrough.csv

This file contains weekly counts and rates of cases and hospitalizations by vaccination status, and citywide. Data are updated on Thursdays containing data through the Saturday 12 days prior. As of 1/6/2022, data from the week of 12/18/2021 onward include children age 5-11 years. Recent data may be incomplete due to delays in processing the surge in cases and linking to vaccination and hospitalization data.  

For more detailed explanation, please check out the link provided above.



- [Link to Covid Genernal Data](https://github.com/nychealth/coronavirus-data/tree/master/trends)

    - data-by-day.csv

This file contains citywide and borough-specific daily counts of probable and confirmed cases. Cases are aggregated by the date of diagnosis. To address variation in the number of cases diagnosed per day, we have included a 7-day average (mean). This is calculated as the average of number of cases on that day and the previous 6 days.
This file includes data since February 29, 2020 based on the date that the Health Department classifies as the start of the COVID-19 outbreak in NYC (i.e., date of the first laboratory-confirmed case).

For more detailed explanation, please check out the link provided above.



- [Link to Shooting Data](https://data.cityofnewyork.us/Public-Safety/NYPD-Shooting-Incident-Data-Year-To-Date-/5ucz-vwe8)

    - NYPD_Shooting_Incident_Data.csv

This is a breakdown of every shooting incident that occurred in NYC during the current calendar year. This data is manually extracted every quarter and reviewed by the Office of Management Analysis and Planning before being posted on the NYPD website.
Each record represents a shooting incident in NYC and includes information about the event, the location and time of occurrence. In addition, information related to suspect and victim demographics is also included. This data can be used by the public to explore the nature of police enforcement activity. Please refer to the attached data footnotes for additional information about this dataset.

For more detailed explanation, please check out the link provided above.



- [Link to NYC Zip Code Boundaries](https://github.com/nychealth/coronavirus-data/tree/master/Geography-resources)

Boundaries for NYC zip codes

For more detailed explanation, please check out the link provided above.



- [Link to Hate Crime Data](https://data.cityofnewyork.us/Public-Safety/NYPD-Hate-Crimes/bqiq-cu78)

    - NYPD_Hate_Crimes.csv

Dataset containing confirmed hate crime incidents in NYC.

For more detailed explanation, please check out the link provided above.
